dalvik-explorer/src/org/jessies/dalvikexplorer/Utils.java: dalvik-explorer/src/org/jessies/dalvikexplorer/DeviceActivity.java: fix CPU speeds that don't fit in a signed 32-bit integer (which tops out around 2GHz).
git-svn-id: http://enh.googlecode.com/svn/trunk@683 825dafb7-1f56-0410-864d-61665d7775f4

+++ dalvik-explorer/src/org/jessies/dalvikexplorer/DeviceActivity.java
@@ -188,8 +188,8 @@ public class DeviceActivity extends TextViewActivity {
     try {
       String minFrequency = Utils.readFile("/sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_min_freq").trim();
       String maxFrequency = Utils.readFile("/sys/devices/system/cpu/cpu0/cpufreq/cpuinfo_max_freq").trim();
-      int minFrequencyHz = Integer.parseInt(minFrequency) * 1000;
-      int maxFrequencyHz = Integer.parseInt(maxFrequency) * 1000;
+      long minFrequencyHz = Long.parseLong(minFrequency) * 1000L;
+      long maxFrequencyHz = Long.parseLong(maxFrequency) * 1000L;
       append(result, "CPU Speed", Utils.prettyHz(maxFrequencyHz) + " (idles at " + Utils.prettyHz(minFrequencyHz) + ")");
     } catch (Exception unexpected) {
       result.append("(Unable to determine CPU frequencies.)");
